server/requirements.txt
flask
flask-security-too
sqlalchemy
stripe
